Move data_sources out of DataItem and into Operation.
Introduce a general DataSource mechanism that can be
constructed hierarchically. OperationItems and the newly
introduced DataItemDataSource are the first classes that
are considered to be DataSources.

Rework the connect_data_sources method to be more
general and easy to understand.

Rework the dependent_data_item techniques too to go along
with the other architectural changes.

Was svn r2567