Add test dialog script (run script).
The idea is to declare the tests in JSON and run them from the
dialog. The JSON test descriptions attempt to be a generic description
language for acquisitions.

The tests change defocus and energy offset and may leave the
microscope or acquisition dialogs in a different state than at start.